two versions of downloader
one uses pytube and the alt version yt_dlp both work for individual videos and playlist of youtube. for now it only has an interface with input and button to download, the next I'll do is include progress and perhaps options to choose quality. Both programs save the videofiles to the root folder of the project.