new project sum-toggl-wroktime-by-tag
I made this python project to make more easy for me to get the complete number of hours worked for each tag I use in the month so I don't need to do the math manually. The code simply takes a complete export as csv of my toggl.com registered data for a period of time (including tags, obviously). Then makes a table with one row for each tag type and the total of hours I spent working in each one. Anyone can try it out, just remember to add the csv file to same folder and change the name of the csv file in the code.

In the future I'll try to add a interface and the possibility to chose the csv file from there and the path where we want to save the result.